requirements.txt
langchain>=0.1.0
langgraph>=0.0.0
langgraph-supervisor>=0.0.0
langchain-openai>=0.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0

# 数据库连接相关（根据实际需要添加）
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0
# mysql-connector-python>=8.0.0